@@ -26,6 +26,10 @@
       "libs/react-patches",
       "jsdom"
     ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/src/encoded/static/components/lib/"
+    ],
     "coveragePathIgnorePatterns": [
       "/node_modules/",
       "/jest/",
@@ -39,15 +43,24 @@
   },
   "devDependencies": {
     "babel-core": "^6.7.7",
+    "babel-eslint": "^7.2.3",
     "babel-jest": "^12.0.1",
     "babel-loader": "^6.2.4",
-    "babel-plugin-transform-object-rest-spread": "^6.6.5",
-    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-syntax-class-properties": "^6.13.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-flow-strip-types": "^6.22.0",
+    "babel-plugin-transform-object-rest-spread": "^6.6.5",
     "babel-preset-es2015": "^6.6.0",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.5.0",
     "codacy-coverage": "^2.0.0",
     "coveralls": "^2.11.12",
+    "eslint": "^3.19.0",
+    "eslint-plugin-class-property": "^1.0.6",
+    "eslint-plugin-flowtype": "^2.34.0",
+    "eslint-plugin-react": "^7.0.1",
+    "flow-bin": "^0.47.0",
+    "flowtype-loader": "^0.2.2",
     "gulp": "^3.9.1",
     "gulp-util": "^3.0.7",
     "jest-cli": "^20.0.4",
@@ -60,45 +73,48 @@
   "dependencies": {
     "auth0-lock": "10.11.0",
     "aws-sdk": "^2.24.0",
-    "babel-polyfill": "^6.7.4",
+    "babel-polyfill": "^6.23.0",
     "brace": "^0.3.0",
     "ckeditor": "file:node_shims/ckeditor",
     "classnames": "^2.2.5",
     "color": "^0.10.1",
-    "create-react-class": "^15.5.2",
+    "create-react-class": "^15.6.0",
     "crypto-js": "^3.1.9-1",
-    "d3": "^4.4.0",
+    "d3": "^4.9.1",
     "dagre-d3": "git+https://github.com/ENCODE-DCC/dagre-d3.git#compound-nodes",
     "domready": "^0.3.0",
     "form-serialize": "^0.6.0",
     "google-analytics": "file:node_shims/google-analytics",
-    "immutable": "^3.7.5",
-    "js-md5": "^0.4.1",
+    "immutable": "^3.8.1",
+    "js-md5": "^0.4.2",
     "jsonwebtoken": "^7.1.9",
     "lodash": "^4.17.4",
+    "lodash.isfinite": "^3.3.2",
     "marked": "^0.3.6",
-    "moment": "^2.15.2",
-    "prop-types": "^15.5.8",
-    "query-string": "^4.1.0",
-    "radium": "^0.18.2",
-    "rc-progress": "^2.0.6",
-    "rc-tabs": "^7.3.1",
-    "react": "^15.3.2",
+    "moment": "^2.18.1",
+    "prop-types": "^15.5.10",
+    "query-string": "^4.3.4",
+    "radium": "^0.18.4",
+    "rc-progress": "^2.1.2",
+    "rc-tabs": "^7.5.0",
+    "react": "^15.6.1",
     "react-bootstrap": "0.31.0",
-    "react-cookie": "^0.4.8",
-    "react-dom": "^15.3.2",
+    "react-cookie": "^0.4.9",
+    "react-dom": "^15.6.1",
     "react-draggable": "^2.2.6",
+    "react-infinite": "^0.11.0",
     "react-json-tree": "^0.10.7",
     "react-matrix-viz": "^0.2.2",
-    "react-redux": "^4.4.5",
-    "react-tooltip": "^3.2.7",
-    "redux": "^3.5.2",
+    "react-redux": "^4.4.8",
+    "react-sticky": "^6.0.1",
+    "react-tooltip": "^3.3.0",
+    "redux": "^3.7.0",
     "remark-parse": "^3.0.0",
-    "scriptjs": "^2.5.7",
-    "source-map-support": "^0.2.7",
+    "scriptjs": "^2.5.8",
+    "source-map-support": "^0.2.10",
     "subprocess-middleware": "^0.1.0",
     "underscore": "^1.8.3",
     "unified": "^6.1.2",
-    "xhr2": "^0.1.3"
+    "xhr2": "^0.1.4"
   }
 }
