@@ -19,12 +19,15 @@ def audit_experiments_have_raw_files(value, system):
     error is generated if files are missing and status is 'released'
     '''
     ok_stati = ['uploaded', 'released', 'in review by project', 'released by project']
+    raw_types = ['fastq']
     files = value.get('files', None)
     has_raw = False
     if files is not None:
         for i in range(len(files)):
             f = value['files'][i]
-            if f['file_classification'] == 'raw file' and f['status'] in ok_stati:
+            fclass = f.get('file_classification', 'unclassified')
+            ftype = f.get('file_type', 'notype')
+            if ((fclass == 'raw file' or ftype in raw_types) and f['status'] in ok_stati):
                 has_raw = True
                 break
     if not has_raw:
