@@ -28,7 +28,7 @@ describe('Testing browse.js for experiment set browser', function() {
         var { Provider, connect } = require('react-redux');
         TestUtils = require('react-dom/lib/ReactTestUtils');
         _ = require('underscore');
-        Browse = require('../browse').Browse;
+        Browse = require('../browse/BrowseView').default;
         context = require('../testdata/browse/context');
         store = require('../../store');
         var dispatch_vals = {
@@ -46,9 +46,16 @@ describe('Testing browse.js for experiment set browser', function() {
         );
     });
 
-    it('has 1 passing entry (an experiment set)', function() {
-        var passEntries = TestUtils.scryRenderedDOMComponentsWithClass(page, 'expset-entry-passed');
-        expect(passEntries.length).toEqual(1);
+    it('has 3 passing entries (experiment sets)', function() {
+        var rows = TestUtils.scryRenderedDOMComponentsWithClass(page, 'search-result-row');
+        var resultRows = rows.filter(function(r){
+            if (r.className.indexOf('fin') > -1) return false;
+            if (r.className.indexOf('empty-block') > -1) return false;
+            if (r.className.indexOf('loading') > -1) return false;
+            return true;
+        });
+
+        expect(resultRows.length).toEqual(context['@graph'].length);
     });
 
     it('filters are rendered correctly (facetlist terms)', function() {
