@@ -11,7 +11,7 @@ jest.dontMock('underscore');
 
 
 describe('Testing viz/QuickInfoBar.js', function() {
-    var React, TestUtils, page, context, filters, _, Wrapper, QuickInfoBar, href;
+    var React, TestUtils, page, context, filters, _, Wrapper, QuickInfoBar, href, expSetFilters;
 
     beforeEach(function() {
         React = require('react');
@@ -21,12 +21,13 @@ describe('Testing viz/QuickInfoBar.js', function() {
         context = require('../testdata/browse/context');
         QuickInfoBar = require('./../viz/QuickInfoBar').default;
         href = "http://localhost:8000/browse/?type=ExperimentSetReplicate&experimentset_type=replicate&limit=25&from=0";
+        expSetFilters = {
+            "experiments_in_set.biosample.biosource.individual.organism.name" : new Set(["mouse"]),
+            "experiments_in_set.biosample.biosource.biosource_type" : new Set(["immortalized cell line"])
+        };
 
         page = TestUtils.renderIntoDocument(
-            <QuickInfoBar href={href} expSetFilters={{
-                "experiments_in_set.biosample.biosource.individual.organism.name" : new Set(["mouse"]),
-                "experiments_in_set.biosample.biosource.biosource_type" : new Set(["immortalized cell line"])
-             }} />
+            <QuickInfoBar href={href} expSetFilters={expSetFilters} />
         );
     });
 
@@ -56,6 +57,29 @@ describe('Testing viz/QuickInfoBar.js', function() {
         statValEls.forEach(function(el){
             expect(el.innerHTML.indexOf('10')).toBeGreaterThan(-1);
             expect(el.innerHTML.indexOf('211')).toBeGreaterThan(-1);
+        });
+    });
+
+    it('There are visible filters in the hover-over dropdown', function() {
+        // Not hovered over yet
+        var crumbs = TestUtils.scryRenderedDOMComponentsWithClass(page, 'chart-crumb');
+        var filterIcon = TestUtils.scryRenderedDOMComponentsWithClass(page, 'glance-label');
+
+        expect(filterIcon.length).toEqual(1);
+        expect(crumbs.length).toEqual(0);
+
+        TestUtils.Simulate.mouseEnter(filterIcon[0]); // Hover over
+
+        var crumbs = TestUtils.scryRenderedDOMComponentsWithClass(page, 'chart-crumb');
+
+        var filterTerms = _.reduce(_.values(expSetFilters), function(m,v){ return m.concat(Array.from(v.values())); }, []);
+
+        console.log("Testing for presence of terms in ActiveFiltersBar: ", filterTerms.join(', '));
+        crumbs.forEach(function(c,i){
+            expect(filterTerms.filter(function(ft){
+                if (c.innerHTML.toLowerCase().indexOf(ft) > -1) return true;
+                return false;
+            }).length).toBeGreaterThan(0);
         });
     });
 
