@@ -10,7 +10,18 @@ jest.dontMock('react');
 jest.dontMock('underscore');
 
 
+function filterRowsToResults(rows){
+    return rows.filter(function(r){
+        if (r.className.indexOf('fin') > -1) return false;
+        if (r.className.indexOf('empty-block') > -1) return false;
+        if (r.className.indexOf('loading') > -1) return false;
+        return true;
+    });
+}
+
+
 describe('Testing search.js', function() {
+    var sinon, server;
     var React, Search, testSearch, TestUtils, context, schemas, _;
 
     beforeEach(function() {
@@ -20,6 +31,19 @@ describe('Testing search.js', function() {
         Search = require('../search').Search;
         context = require('../testdata/expt_search');
 
+        sinon = require('sinon');
+        server = sinon.fakeServer.create();
+
+        server.respondWith(
+            "GET",
+            '/profiles/',
+            [
+                200, 
+                { "Content-Type" : "application/json" },
+                '<html></html>' // Don't actually need content JSON here for test.
+            ]
+        );
+
         testSearch = TestUtils.renderIntoDocument(
             <Search context={context} href="/search/?type=ExperimentHiC" />
         );
@@ -28,7 +52,11 @@ describe('Testing search.js', function() {
 
     it('has the correct number of facets and experiment accessions listed', function() {
         var facets = TestUtils.scryRenderedDOMComponentsWithClass(testSearch, 'facet');
-        var results = TestUtils.scryRenderedDOMComponentsWithClass(testSearch, 'search-result-row');
+        var rows = TestUtils.scryRenderedDOMComponentsWithClass(testSearch, 'search-result-row');
+        var results = filterRowsToResults(rows);
+
+        var facetFields = facets.map(function(f){ return f.getAttribute('data-field'); });
+        console.log("Facets shown for:", facetFields.join(', '));
         expect(facets.length).toBeGreaterThan(7);
         expect(results.length).toEqual(5);
     });
@@ -43,7 +71,8 @@ describe('Testing search.js', function() {
         testSearch = TestUtils.renderIntoDocument(
             <Search context={context} href='/search/?type=ExperimentHiC&digestion_enzyme.name=HindIII' />
         );
-        var results = TestUtils.scryRenderedDOMComponentsWithClass(testSearch, 'search-result-row');
+        var rows = TestUtils.scryRenderedDOMComponentsWithClass(testSearch, 'search-result-row');
+        var results = filterRowsToResults(rows);
         var selectedFacets = TestUtils.scryRenderedDOMComponentsWithClass(testSearch, 'selected');
         expect(results.length).toEqual(2);
         expect(selectedFacets.length).toEqual(2);
