@@ -239,6 +239,7 @@ export default class App extends React.Component {
         // Save navigate fxn and other req'd stuffs to GLOBAL navigate obj.
         // So that we may call it from anywhere if necessary without passing through props.
         navigate.setNavigateFunction(this.navigate);
+        navigate.registerCallbackFunction(Alerts.updateCurrentAlertsTitleMap.bind(this, null));
 
         console.log("App Initial State: ", session, user_actions);
 
@@ -376,7 +377,7 @@ export default class App extends React.Component {
             onDropdownChange: this.handleDropdownChange, // Function to process dropdown state change
             hidePublicAudits: true, // True if audits should be hidden on the UI while logged out
             session: this.state.session,
-            navigate: this.navigate,
+            navigate: navigate,
             schemas : this.state.schemas
         };
     }
@@ -533,7 +534,7 @@ export default class App extends React.Component {
         // through the navigate method.
         if (this.historyEnabled) {
             event.preventDefault();
-            this.navigate(href, navOpts);
+            navigate(href, navOpts);
             if (this.refs && this.refs.navigation){
                 this.refs.navigation.closeMobileMenu();
             }
@@ -571,7 +572,7 @@ export default class App extends React.Component {
 
         if (this.historyEnabled) {
             event.preventDefault();
-            this.navigate(href, options);
+            navigate(href, options);
         }
     }
 
@@ -623,7 +624,7 @@ export default class App extends React.Component {
 
         }
         // Always async update in case of server side changes.
-        this.navigate(href, {'replace': true});
+        navigate(href, {'replace': true});
     }
 
     // If ESC pressed while drop-down menu open, close the menu
@@ -1138,8 +1139,8 @@ export default class App extends React.Component {
                         <SubmissionView
                             {...commonContentViewProps}
                             setIsSubmitting={this.setIsSubmitting}
-                            edit={actionList[0] === 'edit'}
                             create={actionList[0] === 'create'}
+                            edit={actionList[0] === 'edit'}
                         />
                     );
                     title = getTitleStringFromContext(context);
@@ -1236,7 +1237,7 @@ export default class App extends React.Component {
                                         href={this.props.href}
                                         context={this.props.context}
                                         expSetFilters={this.props.expSetFilters}
-                                        navigate={this.navigate}
+                                        navigate={navigate}
                                         updateStats={this.updateStats}
                                         schemas={this.state.schemas}
                                         session={this.state.session}
