@@ -1,5 +1,8 @@
 'use strict';
 
 export { LimitAndPageControls, ColumnSorterIcon } from './LimitAndPageControls';
-export { PageLimitSortController } from './PageLimitSortController';
-export { SearchResultTable } from './SearchResultTable';
+export { SortController } from './SortController';
+export { SelectedFilesController } from './SelectedFilesController';
+export { CustomColumnController, CustomColumnSelector } from './CustomColumnController';
+export { SearchResultTable, defaultColumnBlockRenderFxn, extendColumnDefinitions, defaultColumnDefinitionMap, columnsToColumnDefinitions } from './SearchResultTable';
+export { AboveTableControls } from './AboveTableControls';
