@@ -34,13 +34,13 @@ function setHeaders(xhr, headers = {}) {
 
 
 export function load(url, callback, method = 'GET', fallback = null, data = null, headers = {}){
-    if (typeof window == 'undefined') return null;
+    if (typeof window === 'undefined') return null;
 
     var xhr = new XMLHttpRequest();
     xhr.onreadystatechange = function() {
-        if (xhr.readyState == XMLHttpRequest.DONE ) {
-            if (xhr.status == 200) {
-                if (typeof callback == 'function'){
+        if (xhr.readyState === XMLHttpRequest.DONE ) {
+            if (xhr.status === 200) {
+                if (typeof callback === 'function'){
                     callback(JSON.parse(xhr.responseText));
                 }
             } else {
@@ -48,10 +48,10 @@ export function load(url, callback, method = 'GET', fallback = null, data = null
                 try {
                     response = JSON.parse(xhr.responseText);
                     console.error('ajax.load error: ', response);
-                    if (typeof fallback == 'function') fallback(response);
+                    if (typeof fallback === 'function') fallback(response);
                 } catch (error) {
                     console.error('Non-JSON error response:', xhr.responseText);
-                    if (typeof fallback == 'function') fallback(xhr);
+                    if (typeof fallback === 'function') fallback(xhr);
                 }
             }
         }
