@@ -82,7 +82,7 @@ export function listAllExperimentsFromExperimentSets(experiment_sets){
     var uniqExpAccessions = {};
     return _(experiment_sets).chain()
         .map(function(set){
-            return set.experiments_in_set.map(function(exp){
+            return (set.experiments_in_set || []).map(function(exp){
                 // Make sure we return new exp & set objects instead of mutating existing ones.
                 var cExp = _.clone(exp);
                 var cSet = _.clone(set);
@@ -151,6 +151,8 @@ export function convertToObjectKeyedByAccession(experiments, keepExpObject = tru
         })
     );
 }
+
+
 
 
 /*** *** *** *** *** *** *** *** *** *** *** *** *** *** *** *** *** *** * 
