@@ -8,7 +8,7 @@
  * @return {boolean} - True if processing on serverside.
  */
 export function isServerSide(){
-    if (typeof window == 'undefined' || !window || !window.document || !window.document.createElement){
+    if (typeof window === 'undefined' || !window || !window.document || !window.document.createElement){
         return true;
     }
     return false;
