@@ -13,8 +13,8 @@ export function parseAnalysisSteps(analysis_steps, parsingMethod = 'path'){
 
     /**** Functions ****/
 
-    function stepNodeID   (step) { return step.uuid || step['@id'] || step.link_id; }
-    function stepNodeName (step) { return step.display_title || step.title || step['@id']; }
+    function stepNodeID   (step) { return step.uuid || step['@id'] || step.link_id || step.name; }
+    function stepNodeName (step) { return step.display_title || step.title || step.name || step['@id']; }
 
 
     function generateStepNode(step, column){
