@@ -27,10 +27,10 @@ class Experiment(Item):
     rev = {
         'experiment_sets': ('ExperimentSet', 'experiments_in_set'),
     }
-    embedded = ["protocol", "protocol_variation", "lab", "award", "experiment_sets",
-                "produced_in_pub", "publications_of_exp",
-                "biosample", "biosample.biosource", "biosample.modifications",
-                "biosample.treatments", "biosample.biosource.individual.organism"]
+    embedded = ["protocol.*", "protocol_variation.*", "lab.*", "award.*", "experiment_sets.*",
+                "produced_in_pub.*", "publications_of_exp.*",
+                "biosample.*", "biosample.biosource.*", "biosample.modifications.*",
+                "biosample.treatments.*", "biosample.biosource.individual.organism.*"]
     name_key = 'accession'
 
     def generate_mapid(self, experiment_type, num):
@@ -162,7 +162,7 @@ class ExperimentHiC(Experiment):
 
     item_type = 'experiment_hi_c'
     schema = load_schema('encoded:schemas/experiment_hi_c.json')
-    embedded = Experiment.embedded + ["digestion_enzyme", "submitted_by"]
+    embedded = Experiment.embedded + ["digestion_enzyme.*", "submitted_by.*"]
     name_key = 'accession'
 
     @calculated_property(schema={
@@ -202,11 +202,11 @@ class ExperimentCaptureC(Experiment):
     """The experiment class for Capture Hi-C experiments."""
     item_type = 'experiment_capture_c'
     schema = load_schema('encoded:schemas/experiment_capture_c.json')
-    embedded = Experiment.embedded + ["digestion_enzyme",
-                                      "submitted_by",
-                                      "targeted_regions",
-                                      "targeted_regions.target",
-                                      "targeted_regions.oligo_file"]
+    embedded = Experiment.embedded + ["digestion_enzyme.*",
+                                      "submitted_by.*",
+                                      "targeted_regions.*",
+                                      "targeted_regions.target.*",
+                                      "targeted_regions.oligo_file.*"]
     name_key = 'accession'
 
     @calculated_property(schema={
@@ -247,7 +247,7 @@ class ExperimentRepliseq(Experiment):
     """The experiment class for Repliseq experiments."""
     item_type = 'experiment_repliseq'
     schema = load_schema('encoded:schemas/experiment_repliseq.json')
-    embedded = Experiment.embedded + ["submitted_by"]
+    embedded = Experiment.embedded + ["submitted_by.*"]
     name_key = 'accession'
 
     @calculated_property(schema={
@@ -285,7 +285,7 @@ class ExperimentMic(Experiment):
     """The experiment class for Microscopy experiments."""
     item_type = 'experiment_mic'
     schema = load_schema('encoded:schemas/experiment_mic.json')
-    embedded = Experiment.embedded + ["submitted_by"]
+    embedded = Experiment.embedded + ["submitted_by.*"]
     name_key = 'accession'
 
     @calculated_property(schema={
