@@ -67,7 +67,7 @@ class User(Item):
     item_type = 'user'
     schema = load_schema('encoded:schemas/user.json')
     # Avoid access_keys reverse link so editing access keys does not reindex content.
-    embedded = ['lab.awards']
+    embedded = ['lab.awards.*']
 
     STATUS_ACL = {
         'current': ONLY_OWNER_EDIT,
@@ -75,7 +75,6 @@ class User(Item):
         'replaced': USER_DELETED,
         'revoked': ONLY_ADMIN_VIEW_DETAILS,
     }
-
 
     @calculated_property(schema={
         "title": "Title",
